old-tests/submission/W3C/domhtml/HTMLObjectElement05.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-object-codebase" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<object codebase="http://www.w3c-test.org/" data="/support/html/sample.html" height="80"  type="text/html" width="550"></object>

<script>
var domtest = async_test("The codeBase IDL attribute of the object element must reflect the element's codebase content attribute, which for the purposes of reflection is defined as containing a URL.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcodebase;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcodebase = testNode.codeBase;

      assert_equals(vcodebase,"http://www.w3c-test.org/");
       
    domtest.done();
}));

</script>
<p>The codeBase IDL attribute of the object element must reflect the element's codebase content attribute, which for the purposes of reflection is defined as containing a URL.</p>
<div id="log"></div>
</body></html>
